refactor(stfts_pipeline): slightly revamped the method of reducing non-seizure files
In this change, initially removing non-seizure files was done regardless
of the duration of each patient's recordings, so every patient's
non-seizure recordings will be croppped. But now patients that have an
average recording time of more than 1hr will be only croppped. This is
done to keep data of the patients that have lower recording average
duration